made the use of codegen.List more general
