+++ sympy/codegen/pynodes.py
@@ -1,10 +1,1 @@
-from sympy.core import Tuple
-
-
-class List(Tuple):
-    """Represents a (frozen) (Python) list (for code printing purposes)."""
-    def __eq__(self, other):
-        if isinstance(other, list):
-            return self == List(*other)
-        else:
-            return self.args == other
+from .abstract_nodes import List
