+++ sympy/printing/tests/test_maple.py
@@ -1,3 +1,4 @@
+from sympy.codegen.abstract_nodes import List
 from sympy.core import (S, pi, oo, symbols, Function, Rational, Integer,
                         Tuple, Symbol, Eq, Ne, Le, Lt, Gt, Ge)
 from sympy.core import EulerGamma, GoldenRatio, Catalan, Lambda, Mul, Pow
@@ -233,7 +234,6 @@ def test_containers():
     assert maple_code(Tuple(*[1, 2, 3])) == "[1, 2, 3]"
     assert maple_code((1, x * y, (3, x ** 2))) == "[1, x*y, [3, x^2]]"
     # scalar, matrix, empty matrix and empty list
-    from sympy.codegen.pynodes import List
     assert maple_code((1, eye(3), Matrix(0, 0, []), List(*[]))) == \
            "[1, Matrix([[1, 0, 0], [0, 1, 0], [0, 0, 1]], storage = rectangular), Matrix([], storage = rectangular), []]"
