+++ sympy/printing/tests/test_octave.py
@@ -1,3 +1,4 @@
+from sympy.codegen.abstract_nodes import List
 from sympy.core import (S, pi, oo, symbols, Function, Rational, Integer,
                         Tuple, Symbol, EulerGamma, GoldenRatio, Catalan,
                         Lambda, Mul, Pow, Mod, Eq, Ne, Le, Lt, Gt, Ge)
@@ -281,7 +282,6 @@ def test_containers():
     assert mcode(Tuple(*[1, 2, 3])) == "{1, 2, 3}"
     assert mcode((1, x*y, (3, x**2))) == "{1, x.*y, {3, x.^2}}"
     # scalar, matrix, empty matrix and empty list
-    from sympy.codegen.pynodes import List
     assert mcode((1, eye(3), Matrix(0, 0, []), List(*[]))) == "{1, [1 0 0; 0 1 0; 0 0 1], [], {}}"
